pinvoke GetInteractiveUserList() change int to Int64
- I've recycled this function in another project and received Exeptions in Win10
  moving to int64 fixed it

+++ pGina/src/Abstractions/WindowsApi/pInvokes.cs
@@ -1103,7 +1103,7 @@ namespace Abstractions.WindowsApi
             if(retVal != 0)
             {
                 int dataSize = Marshal.SizeOf(typeof(SafeNativeMethods.WTS_SESSION_INFO));
-                int currentSession = (int) sessionInfoList;
+                Int64 currentSession = (Int64) sessionInfoList;
 
                 for(int x = 0; x < sessionCount; x++)
                 {
